Mobile Application Protype
This is the bas skeleton for our mobile application.

There are 5 buttons to test.

Please use the LED button to test connectivity to the hardware.

Push and commit bugged versions of the file as a different name (CollarCompanionV1Bugged).

Working changes can be kept in the same file (CollarCompanion).

If any further explanation is needed, please contact Cody.

Run this folder in Android Studios.